+++ test/analyzer.js
@@ -114,6 +114,35 @@ describe('Analyzer', function () {
     const chartData = require(path.resolve(__dirname, 'output/report.json'));
     expect(chartData).to.containSubset(require('./stats/with-modules-in-chunks/expected-chart-data'));
   });
+
+  describe('options', function () {
+    describe('title', function () {
+      it('should take the --title option', async function () {
+        const reportTitle = 'A string report title';
+        generateReportFrom('with-modules-chunk.json', `--title "${reportTitle}"`);
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.equal(reportTitle);
+      });
+      it('should take the -t option', async function () {
+        const reportTitle = 'A string report title';
+
+        generateReportFrom('with-modules-chunk.json', `-t "${reportTitle}"`);
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.equal(reportTitle);
+      });
+      it('should use a suitable default title', async function () {
+        generateReportFrom('with-modules-chunk.json');
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.match(/^webpack-bundle-analyzer \[.* at \d{2}:\d{2}\]/u);
+      });
+    });
+  });
 });
 
 function generateJSONReportFrom(statsFilename) {
@@ -122,10 +151,16 @@ function generateJSONReportFrom(statsFilename) {
   });
 }
 
-function generateReportFrom(statsFilename) {
-  childProcess.execSync(`../lib/bin/analyzer.js -m static -r output/report.html -O stats/${statsFilename}`, {
-    cwd: __dirname
-  });
+function generateReportFrom(statsFilename, additionalOptions = '') {
+  childProcess.execSync(
+    `../lib/bin/analyzer.js ${additionalOptions} -m static -r output/report.html -O stats/${statsFilename}`,
+    {
+      cwd: __dirname
+    });
+}
+
+async function getTitleFromReport() {
+  return await nightmare.goto(`file://${__dirname}/output/report.html`).title();
 }
 
 async function getChartData() {