+++ test/plugin.js
@@ -96,6 +96,63 @@ describe('Plugin', function () {
         expect(_.map(chartData, 'label')).to.deep.equal(['bundle.js']);
       });
     });
+
+    describe('reportTitle', function () {
+      it('should have a sensible default', async function () {
+        const config = makeWebpackConfig();
+
+        await webpackCompile(config);
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.match(/^webpack-bundle-analyzer \[.* at \d{2}:\d{2}\]/u);
+      });
+      it('should use a string', async function () {
+        const reportTitle = 'A string report title';
+        const config = makeWebpackConfig({
+          analyzerOpts: {
+            reportTitle
+          }
+        });
+
+        await webpackCompile(config);
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.equal(reportTitle);
+      });
+      it('should use a function', async function () {
+        const reportTitleResult = 'A string report title';
+        const config = makeWebpackConfig({
+          analyzerOpts: {
+            reportTitle: () => reportTitleResult
+          }
+        });
+
+        await webpackCompile(config);
+
+        const generatedReportTitle = await getTitleFromReport();
+
+        expect(generatedReportTitle).to.equal(reportTitleResult);
+      });
+      it('should propogate an error in a function', async function () {
+        const reportTitleError = new Error();
+        const config = makeWebpackConfig({
+          analyzerOpts: {
+            reportTitle: () => {throw reportTitleError}
+          }
+        });
+
+        let error = null;
+        try {
+          await webpackCompile(config);
+        } catch (e) {
+          error = e;
+        }
+
+        expect(error).to.equal(reportTitleError);
+      });
+    });
   });
 
   async function expectValidReport(opts) {
@@ -121,6 +178,10 @@ describe('Plugin', function () {
 
   function getChartDataFromJSONReport(reportFilename = 'report.json') {
     return require(path.resolve(__dirname, `output/${reportFilename}`));
+  }
+
+  async function getTitleFromReport(reportFilename = 'report.html') {
+    return await nightmare.goto(`file://${__dirname}/output/${reportFilename}`).title();
   }
 
   async function getChartDataFromReport(reportFilename = 'report.html') {